+++ src/main.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App'
+import CloudApp from './CloudApp'
 import SettingsCenter from './SettingsCenter'
 import AdaptiveNav from './AdaptiveNav'
 import './styles.css'
 
 createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <CloudApp />
     <SettingsCenter />
     <AdaptiveNav />
   </React.StrictMode>,